requirements-langserve.txt
langgraph>=0.0.66
langchain_core>=0.2.5
langchain_openai>=0.1.8
langchain_community>=0.2.4
python-dotenv>=1.0.1
requests>=2.32.3
langchain_experimental>=0.0.60



#for Tool
numexpr
# tavily-python
pandas

#vector db
# chromadb
langchain_chroma
chromadb==0.5.3

wikipedia
#langgraph AsyncSqliteSaver
aiosqlite

langserve[all]

pydantic==1.10.13

tavily-python